refactor: consolidated alignment stages in align/, refactored monolithic eval/services.py
