+++ src/prml_vslam/align/gravity/__init__.py
@@ -1,0 +1,15 @@
+"""Ground-alignment pipeline stage integration."""
+
+from .config import GroundAlignmentStageConfig
+from .contracts import GroundAlignmentConfig
+from .runtime import GroundAlignmentRuntime
+from .services import GroundAlignmentService
+from .stage_contracts import GroundAlignmentStageInput
+
+__all__ = [
+    "GroundAlignmentConfig",
+    "GroundAlignmentRuntime",
+    "GroundAlignmentService",
+    "GroundAlignmentStageConfig",
+    "GroundAlignmentStageInput",
+]
