+++ src/prml_vslam/align/icp/__init__.py
@@ -1,0 +1,5 @@
+"""ICP point-cloud alignment — algorithm and pipeline stage."""
+
+from ._algorithm import CloudAlignmentService
+
+__all__ = ["CloudAlignmentService"]
