+++ src/prml_vslam/align/icp/stage_contracts.py
