+++ src/prml_vslam/align/trajectory_sim3/__init__.py
@@ -1,0 +1,10 @@
+"""Sim(3) trajectory alignment — algorithm and pipeline stage."""
+
+from ._algorithm import align_estimate_sim3, is_gravity_aligned_target, sim3_up_axis_tilt_deg, trajectory_supports_sim3
+
+__all__ = [
+    "align_estimate_sim3",
+    "is_gravity_aligned_target",
+    "sim3_up_axis_tilt_deg",
+    "trajectory_supports_sim3",
+]
