+++ src/prml_vslam/align/trajectory_sim3/contracts.py
