+++ src/prml_vslam/align/trajectory_sim3/stage_contracts.py
