--- src/prml_vslam/alignment/__init__.py
@@ -1,9 +1,0 @@
-"""Derived alignment contracts and services."""
-
-from .contracts import GroundAlignmentConfig
-from .services import GroundAlignmentService
-
-__all__ = [
-    "GroundAlignmentConfig",
-    "GroundAlignmentService",
-]
