--- src/prml_vslam/alignment/stage/__init__.py
@@ -1,13 +1,0 @@
-"""Ground-alignment pipeline stage integration."""
-
-from __future__ import annotations
-
-from prml_vslam.alignment.stage.config import GroundAlignmentStageConfig
-from prml_vslam.alignment.stage.contracts import GroundAlignmentStageInput
-from prml_vslam.alignment.stage.runtime import GroundAlignmentRuntime
-
-__all__ = [
-    "GroundAlignmentRuntime",
-    "GroundAlignmentStageInput",
-    "GroundAlignmentStageConfig",
-]
