+++ src/prml_vslam/eval/services/__init__.py
@@ -1,0 +1,10 @@
+"""Evaluation services: trajectory APE/RPE computation and persistence."""
+
+from .preview import compute_trajectory_ape_preview, compute_trajectory_rpe_preview
+from .trajectory_evaluation import TrajectoryEvaluationService
+
+__all__ = [
+    "TrajectoryEvaluationService",
+    "compute_trajectory_ape_preview",
+    "compute_trajectory_rpe_preview",
+]
