--- src/prml_vslam/eval/stage_alignment/__init__.py
@@ -1,1 +1,0 @@
-"""Trajectory alignment stage — Sim(3) alignment without APE metrics."""
