--- src/prml_vslam/eval/stage_cloud_alignment/__init__.py
@@ -1,13 +1,0 @@
-"""Offline point-cloud alignment stage exports."""
-
-from .config import CloudAlignmentStageConfig
-from .contracts import CloudAlignmentStageInput
-from .runtime import CloudAlignmentRuntime
-from .spec import CLOUD_ALIGNMENT_STAGE_SPEC
-
-__all__ = [
-    "CLOUD_ALIGNMENT_STAGE_SPEC",
-    "CloudAlignmentRuntime",
-    "CloudAlignmentStageConfig",
-    "CloudAlignmentStageInput",
-]
