+++ src/prml_vslam/pipeline/config.py
@@ -11,10 +11,10 @@ from typing import Annotated, Any, Literal, Self, TypeAlias, Union, get_args, ge
 
 from pydantic import BaseModel, ConfigDict, Field, PrivateAttr, model_validator
 
-from prml_vslam.alignment.stage.config import GroundAlignmentStageConfig
-from prml_vslam.eval.stage_alignment.config import TrajectoryAlignmentStageConfig
+from prml_vslam.align.gravity.config import GroundAlignmentStageConfig
+from prml_vslam.align.icp.config import CloudAlignmentStageConfig
+from prml_vslam.align.trajectory_sim3.config import TrajectoryAlignmentStageConfig
 from prml_vslam.eval.stage_cloud.config import CloudEvaluationStageConfig
-from prml_vslam.eval.stage_cloud_alignment.config import CloudAlignmentStageConfig
 from prml_vslam.eval.stage_trajectory.config import (
     TrajectoryEvaluationPolicy,
     TrajectoryEvaluationStageConfig,
